requirements.txt
PyYAML
sqlalchemy==1.3.11
mysql-connector-python==8.0.18
mysqlclient==1.4.6
pip-chill==0.1.8
spacy==2.2.2
jupyter==1.0.0
pymongo==3.9.0
pandas==0.25.3
